temporary commit

+++ src/main/java/indi/etern/checkIn/action/role/DeleteRoleAction.java
@@ -1,20 +1,23 @@
 package indi.etern.checkIn.action.role;
 
-import indi.etern.checkIn.action.TransactionalAction;
+import indi.etern.checkIn.action.BaseAction;
+import indi.etern.checkIn.action.MessageOutput;
 import indi.etern.checkIn.action.interfaces.Action;
+import indi.etern.checkIn.action.interfaces.ExecuteContext;
+import indi.etern.checkIn.action.interfaces.InputData;
 import indi.etern.checkIn.entities.user.Role;
 import indi.etern.checkIn.service.dao.RoleService;
 import indi.etern.checkIn.service.web.WebSocketService;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Optional;
 
 @Action("deleteRole")
-public class DeleteRoleAction extends TransactionalAction {
+public class DeleteRoleAction extends BaseAction<DeleteRoleAction.Input, MessageOutput> {
+    public record Input(String roleType) implements InputData {}
+    
     private final RoleService roleService;
     private final WebSocketService webSocketService;
-    String roleType;
     
     public DeleteRoleAction(RoleService roleService, WebSocketService webSocketService) {
         super();
@@ -23,31 +26,24 @@ public class DeleteRoleAction extends TransactionalAction {
     }
     
     @Override
-    public String requiredPermissionName() {
-        return "delete role";
-    }
-    
-    @Override
-    protected Optional<LinkedHashMap<String,Object>> doAction() throws Exception {
-        Role role = roleService.findByType(roleType).orElseThrow();
+    @Transactional
+    public void execute(ExecuteContext<DeleteRoleAction.Input, MessageOutput> context) {
+        context.requirePermission("delete role");
+        Role role = roleService.findByType(context.getInput().roleType).orElseThrow();
         if (!role.getUsers().isEmpty()) {
-            return getOptionalErrorMap("用户组非空");
+            context.resolve(MessageOutput.error("Role is not empty"));
+        } else {
+            roleService.delete(role);
+            
+            LinkedHashMap<String,Object> map = new LinkedHashMap<>();
+            map.put("type","deleteRole");
+            LinkedHashMap<String,Object> role1 = new LinkedHashMap<>();
+            role1.put("type",role.getType());
+            role1.put("level",role.getLevel());
+            map.put("role",role1);
+            webSocketService.sendMessageToAll(map);
+            
+            context.resolve(MessageOutput.success("Role deleted"));
         }
-        roleService.delete(role);
-        
-        LinkedHashMap<String,Object> map = new LinkedHashMap<>();
-        map.put("type","deleteRole");
-        LinkedHashMap<String,Object> role1 = new LinkedHashMap<>();
-        role1.put("type",role.getType());
-        role1.put("level",role.getLevel());
-        map.put("role",role1);
-        webSocketService.sendMessageToAll(map);
-        
-        return successOptionalMap;
-    }
-
-    @Override
-    public void initData(Map<String, Object> dataMap) {
-        roleType = (String) dataMap.get("roleType");
     }
 }
